### 2025-07-04 #715 @ScottClaessens
coevolve - presubmission inquiry
Submitting Author Name: Scott Claessens
Submitting Author Github Handle: @ScottClaessens
Repository:  https://github.com/ScottClaessens/coevolve
Submission type: Pre-submission

---

```
Package: coevolve
Title: Fit Bayesian Generalized Dynamic Phylogenetic Models using 'Stan'
Version: 0.1.0.9002
    c(person("Scott", "Claessens", , "scott.claessens@gmail.com",
             role = c("aut", "cre"), comment = c(ORCID = "0000-0002-3562-6981")),
      person("Erik", "Ringen", , "erikjacob.ringen@uzh.ch",
             role = c("aut")))
Description: Fit Bayesian generalized dynamic phylogenetic models using 'Stan'.
    The package allows an abritrary number of variables with different response
    distributions to coevolve on a phylogenetic tree via an Ornstein-Uhlenbeck
    evolutionary process, allowing users to assess directionality between
    coevolving variables.
    withr
    ape,
    bayesplot,
    cmdstanr,
    colorspace,
    Matrix,
    patchwork,
    phangorn,
    phaseR,
    phytools,
    posterior,
    tidyr
Additional_repositories: https://stan-dev.r-universe.dev/
Remotes: stan-dev/cmdstanr
```

The package is designed to fit Bayesian generalized dynamic phylogenetic models in Stan. The package writes the Stan code, prepares the Stan data list, and fits the model with cmdstanr.

Not yet, but I'm in the process of doing this. I have chosen the "bayesian" category - let me know if you think this most appropriate.

The target audience is researchers in the life sciences, cultural evolution, etc. who would like to model how multiple variables have coevolved over evolutionary time.

To my knowledge, this is the only R package that implements generalized dynamic phylogenetic models in Stan. The software BayesTraits achieves something similar (https://www.evolution.reading.ac.uk/BayesTraitsV3/BayesTraitsV3.html) but it is limited to binary traits.

I don't think these guidelines apply to this package.

I have already endeavoured to meet the rOpenSci requirements (e.g., pkgcheck, lintr, codecov). Unfortunately, I've been having issues testing the package on my Windows machine (see [here](https://github.com/ropensci-review-tools/pkgstats/issues/65)) and the pkgcheck GitHub action isn't working for me (see [here](https://github.com/ropensci-review-tools/pkgcheck-action/issues/54)). I would appreciate some help getting these checks over the line!
