### 2025-07-07 #715 @ScottClaessens
Thanks for the quick response @mpadge!

Great, I will try and get `pkgcheck` working locally with those changes.

From eyeballing the general and category-specific (Bayesian) standards, it looks like it will possible to comply with at least half. I'll continue to work through them.

I took a look at the EDA standards. As well as various visualisation functions which are dedicated to understanding the implications of the fitted model, we also have a `summary` method that prints a summary of the model parameters (similar to [brms](https://paulbuerkner.com/brms/) output). So it seems like these standards will apply. I will incorporate these standards into the compliance process.
