### 2025-07-07 79fd4a5
Initial compliance checks with srr
