### 2025-07-07 37ba231
Tests for missing or undefined values in function returns
