### 2025-07-12 ef71a39
Add plot method for coevfit class
