+++ tests/testthat/test-coev_fit.R
@@ -1,6 +1,3 @@
-# flag for extended tests
-run_extended_tests <- identical(Sys.getenv("COEVOLVE_EXTENDED_TESTS"), "true")
-
 test_that("coev_fit() produces expected errors", {
   # simulate data
   withr::with_seed(1, {
@@ -30,6 +27,7 @@ test_that("coev_fit() produces expected errors", {
     "Argument 'data' must be coercible to a data.frame.",
     fixed = TRUE
   )
+  #' @srrstats {G5.8, G5.8a} Test for zero-length data
   expect_error(
     coev_fit(
       data = data.frame(), # empty
@@ -98,6 +96,59 @@ test_that("coev_fit() produces expected errors", {
     "Must be at least two coevolving variables.",
     fixed = TRUE
   )
+  expect_error(
+    coev_fit(
+      data = dplyr::tibble(
+        id = tree$tip.label,
+        x = list("test"),
+        y = list("test")
+      ),
+      variables = list(
+        x = "bernoulli_logit",
+        y = "ordered_logistic"
+      ),
+      id = "id",
+      tree = tree
+    ),
+    "Data must not contain list columns.",
+    fixed = TRUE
+  )
+  expect_error(
+    coev_fit(
+      data = data.frame(
+        id = tree$tip.label,
+        x = Inf,
+        y = -Inf
+      ),
+      variables = list(
+        x = "bernoulli_logit",
+        y = "ordered_logistic"
+      ),
+      id = "id",
+      tree = tree
+    ),
+    "Data must not contain undefined values (i.e., Inf or -Inf).",
+    fixed = TRUE
+  )
+  #' @srrstats {G5.8, G5.8c} Test for columns with only NA values
+  expect_error(
+    coev_fit(
+      data = data.frame(
+        id = tree$tip.label,
+        x = NA,
+        y = NA
+      ),
+      variables = list(
+        x = "bernoulli_logit",
+        y = "ordered_logistic"
+      ),
+      id = "id",
+      tree = tree
+    ),
+    "Data must not contain columns with only NA values.",
+    fixed = TRUE
+  )
+  #' @srrstats {G5.8, G5.8b, G5.8d} Tests for data of unsupported types
   expect_error(
     coev_fit(
       data = d,
@@ -716,105 +767,45 @@ test_that("coev_fit() produces expected errors", {
   )
 })
 
-test_that("coev_fit() fits simple model without error", {
-  # simulate data
-  withr::with_seed(1, {
-    n <- 3
-    tree <- ape::rcoal(n)
-    d <- data.frame(
-      id = tree$tip.label,
-      x = rnorm(n),
-      y = rnorm(n)
-    )
-  })
-  # fit model
-  fit <-
-    coev_fit(
-      data = d,
-      variables = list(
-        x = "normal",
-        y = "normal"
-      ),
-      id = "id",
-      tree = tree,
-      chains = 1,
-      seed = 1,
-      refresh = 0
-    )
-  # expect no errors for model fitting or summaries
-  sw <- suppressWarnings
-  expect_no_error(sw(fit))
-  expect_no_error(sw(summary(fit)))
-  expect_no_error(sw(print(fit)))
-  expect_no_error(sw(print(summary(fit))))
-  # expect no error for stancode and standata methods
-  expect_no_error(sw(stancode(fit)))
-  expect_no_error(sw(standata(fit)))
-  expect_output(sw(stancode(fit)))
-  expect_true(sw(methods::is(standata(fit), "list")))
-})
+# test_that("coev_fit() fits simple model without error", {
+#   # simulate data
+#   withr::with_seed(1, {
+#     n <- 3
[…]
