+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(extract_samples,coevfit)
+S3method(plot,coevfit)
 S3method(print,coevfit)
 S3method(print,coevsummary)
 S3method(stancode,coevfit)
