### 2025-07-12 a4bc961
Implement initial Bayesian compliance checks
