### 2025-07-12 #715 @ScottClaessens
I have finished implementing the `srr` standards for this package. In the end, I decide not to implement the EDA standards and focus solely on the general and Bayesian-specific standards - I hope that's okay!

As I mentioned [here](https://github.com/ropensci-review-tools/pkgstats/pull/80#issuecomment-3065795762), I was able to get the `pkgcheck()` function working locally on my Windows machine. Here is the output:

```
── coevolve 0.1.0.9002 ───────────────────────────────────────────────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✖ Package has no continuous integration checks.
✔ Package coverage is 98.3%.
✔ This is a statistical package which complies with all applicable standards
✔ R CMD check found no errors.
✔ R CMD check found no warnings.
ℹ Package has unusually large number of 19 Imports (> 97% of all packages)

ℹ Current status:
✖ This package is not ready to be submitted.

── git ──

• HEAD: 532baccd
• Default branch: main
• Number of commits: 241
• First commit: 18-10-2023
• Number of authors: 4

── rOpenSci Statistical Standards ──

ℹ The package is in the following 1 category:
• Bayesian and Monte Carlo

ℹ Compliance with rOpenSci statistical standards:
✔ All applicable standards [v0.2.0] have been documented in this package (470 complied with; 24 N/A standards)
ℹ 'srr' report is at [C:/Users/sc2223/AppData/Local/Cache/R/pkgcheck/static/coevolve_srr532baccd.html].

── Package Structure ──

ℹ Package uses the following languages:
• R: 100%

ℹ Package has
• 2 authors.
• 1 vignette.
• 3 internal data files.
• 19 imported packages.
• 21 exported functions (median 75 lines of code).
• 82 non-exported functions (median 37 lines of code).
• 4 parameters per function (median).

── Package statistics ──
                    measure   value percentile noteworthy
1                   files_R    23.0       83.6
4           files_vignettes     2.0       81.8
5               files_tests    17.0       93.3
6                     loc_R  3701.0       91.3
7             loc_vignettes   703.0       84.2
8                 loc_tests  5364.0       97.9       TRUE
9             num_vignettes     1.0       58.4
10          data_size_total 67929.0       80.3
11         data_size_median  8815.0       77.9
12                  n_fns_r   103.0       75.6
13         n_fns_r_exported    21.0       68.4
14     n_fns_r_not_exported    82.0       77.8
15         n_fns_per_file_r     2.8       47.6
16        num_params_per_fn     4.0       51.2
17             loc_per_fn_r    41.0       85.4
18         loc_per_fn_r_exp    75.0       87.9
19     loc_per_fn_r_not_exp    37.0       83.9
20         rel_whitespace_R     0.9       23.9
21 rel_whitespace_vignettes    24.3       78.4
22     rel_whitespace_tests     0.8       49.7
23      doclines_per_fn_exp    69.0       78.7
24  doclines_per_fn_not_exp     0.0        0.0       TRUE
25     fn_call_network_size   249.0       89.9
```

@mpadge Please can you help me understand the following:

1. Why is `pkgcheck()` saying that the package has no continuous integration tests? We have implemented GitHub actions for R CMD checks, lintr checks, and code coverage.
2. Why do the package statistics report that we have no documentation for non-exported functions? I made sure to go through the package and write documentation for non-exported functions (e.g., [here](https://github.com/ScottClaessens/coevolve/blob/532baccdf8f41e746e9a6f4ed75e71d1382d7eeb/R/coev_make_stancode.R#L332-L341)).
