+++ DESCRIPTION
@@ -18,7 +18,6 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
 RoxygenNote: 7.3.2
 Suggests: 
-    gridExtra,
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
@@ -32,6 +31,7 @@ Imports:
     colorspace,
     dplyr,
     ggplot2,
+    gridExtra,
     Matrix,
     methods,
     patchwork,
