### 2025-07-13 #715 @adamhsparks
Hi @ScottClaessens, @mpadge is away for a bit and I'm filling in for him as EIC.

I'm running checks on {coevolve} locally using an Ubuntu 2504 image and have found two easy to fix errors:

```r
N  checking top-level files ...
   Non-standard file/directory found at top level:
     ‘codemeta.json’
```

and

```r
N  checking dependencies in R code (894ms)
   Namespace in Imports field not imported from: ‘gridExtra’
     All declared Imports should be used.
```

Regarding your question about CI checks. It appears to be a bug in {pkgcheck}. I've tested it against some of my own rOpenSci packages that also have CI and it doesn't pick them up. We'll have to wait for Mark to return to work on that issue.

Regarding the documentation for non-exported functions, I'm unsure on this one as well, Mark will have to field this question too as I'm not intimately familiar with {pkgstats} internals but from looking at your package, it looks correct to me.
