+++ R/coevolve-package.R
@@ -1,7 +1,7 @@
 #' @keywords internal
 #' @importFrom rlang .data
+#' @importFrom gridExtra grid.arrange
 "_PACKAGE"
-
 
 ## usethis namespace: start
 ## usethis namespace: end
