+++ NAMESPACE
@@ -24,4 +24,5 @@ export(extract_samples)
 export(save_coevfit)
 export(stancode)
 export(standata)
+importFrom(gridExtra,grid.arrange)
 importFrom(rlang,.data)
