### 2025-07-14 #715 @ScottClaessens
Thanks @adamhsparks for checking the package.

Regarding the first note, one of the `pkgcheck` criteria is that the package should include a `codemeta.json` file. The `codemetar::write_codemeta()` function automatically writes this file at the top level of the directory. Should I move this file elsewhere?

I have fixed the second note in the most recent commit.
