### 2025-07-15 e3c88da
Fix failing extended tests
