+++ tests/testthat/test-extended.R
@@ -36,6 +36,8 @@ test_that("coev_fit() estimates correct direction of selection", {
       seed = 1,
       refresh = 0
     )
+  # check model converged
+  expect_lt(max(suppressWarnings(fit$fit$summary())$rhat, na.rm = TRUE), 1.1)
   # get delta theta values
   delta_theta_x_to_y <-
     coev_calculate_delta_theta(
@@ -69,7 +71,7 @@ for (seed in 1:3) {
     skip_if_not(run_extended_tests)
     # get dummy data
     withr::with_seed(seed, {
-      n <- 60
+      n <- 100
       tree <- ape::rcoal(n)
       d <- data.frame(
         id = tree$tip.label,
@@ -102,16 +104,19 @@ for (seed in 1:3) {
     )
     scode_fixed <- manually_fix_parameters(scode)
     # simulate dataset from model with fixed parameters
-    sim <- cmdstanr::cmdstan_model(
-      stan_file = cmdstanr::write_stan_file(scode_fixed)
-    )$sample(
-      data = sdata,
-      chains = 1,
-      refresh = 0,
-      seed = seed,
-      iter_warmup = 50,
-      iter_sampling = 1
-    )
+    sim <-
+      suppressWarnings(
+        cmdstanr::cmdstan_model(
+          stan_file = cmdstanr::write_stan_file(scode_fixed)
+        )$sample(
+          data = sdata,
+          chains = 1,
+          refresh = 0,
+          seed = seed,
+          iter_warmup = 50,
+          iter_sampling = 1
+        )
+      )
     draws <- posterior::as_draws_rvars(sim)
     d_sim <- data.frame(
       id = tree$tip.label,
@@ -122,41 +127,46 @@ for (seed in 1:3) {
     effects_mat <- matrix(TRUE, nrow = 2, ncol = 2,
                           dimnames = list(c("x", "y"), c("x", "y")))
     effects_mat[1, 2] <- FALSE
-    fit <- coev_fit(
-      data = d_sim,
-      variables = list(
-        x = "normal",
-        y = "normal"
-      ),
-      id = "id",
-      tree = tree,
-      estimate_correlated_drift = FALSE,
-      effects_mat = effects_mat,
-      prior = list(
-        A_offdiag = "normal(0, 2)",
-        Q_sigma = "normal(0, 2)"
-      ),
-      scale = FALSE,
-      chains = 1,
-      refresh = 0,
-      seed = seed,
-      max_treedepth = 15
-    )
-    post <- extract_samples(fit)
+    fit <-
+      suppressWarnings(
+        coev_fit(
+          data = d_sim,
+          variables = list(
+            x = "normal",
+            y = "normal"
+          ),
+          id = "id",
+          tree = tree,
+          estimate_correlated_drift = FALSE,
+          effects_mat = effects_mat,
+          prior = list(
+            A_offdiag = "normal(0, 2)",
+            Q_sigma = "normal(0, 2)"
+          ),
+          scale = FALSE,
+          chains = 1,
+          refresh = 0,
+          seed = seed,
+          max_treedepth = 15
+        )
+      )
+    # check model converged
+    expect_lt(max(suppressWarnings(fit$fit$summary())$rhat, na.rm = TRUE), 1.1)
     # posterior medians should recover fixed parameters within tolerance
-    expect_equal(median(post$A[, 1, 1]), -0.5, tolerance = 0.5)
[…]
