### 2025-07-18 #715 @mpadge
Thanks @ScottClaessens, that's great news, and yes it is indeed perfectly okay to document compliance with Bayesian and Monte Carlo standards only. I think your `pkgcheck` issue with continuous integration is because you're using an older version; if you update to latest, that should be fine (version >= 0.1.2.194).

The statistics aspect that non-exported functions have no doclines is indeed a bug, and will be addressed via `pkgstats` issue linked above. Thanks for that! One that's fixed, you'll see the green circle turn purple for closed issue. If you then install latest `pkgstats` and `pkgcheck`, everything should work as expected. Feel free then to call `@ropensci-review-bot check package` here to confirm.
