### 2025-07-18 #715 @adamhsparks
Sorry, @ScottClaessens, I’d missed your question. The codemeta.jaon is where it should be. But it should be ignored in the .Rbuildignore file.
