### 2025-07-22 #715 @ScottClaessens
Hi @mpadge @adamhsparks, the doclines bug issue hasn't closed yet, but I thought I would try running `pkgcheck()` anyway. The coevolve package now passes all main checks.
