### 2025-07-23 #715 @mpadge
@ScottClaessens Sorry about the glitches with out bot checks there. You'll see the biggest issue was reported just before the checks: You list a dependency on `phaseR`, but that package [was archived on 22-07-2025](https://cran.r-project.org/web/packages/phaseR/index.html) - yesterday :sob:. Easiest fix for the moment would be to install from https://github.com/mjg211/phaseR by adding that as an additional `Remotes` in your DESCRIPTION file.
