### 2025-07-23 d4e4b61
Fix DESCRIPTION formatting
